FieldSettings - split file into more
