+++ projects/eav-ui/src/app/edit/fields/basic/custom-json-editor/custom-json-editor-logic.ts
@@ -1,4 +1,5 @@
-import { CustomJsonEditor, FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
+import { CustomJsonEditor } from 'projects/edit-types/src/FieldSettings-CustomJsonEditor';
+import { FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
 import { InputTypeCatalog } from '../../../../shared/fields/input-type-catalog';
 import { FieldLogicBase, FieldLogicUpdate } from '../../logic/field-logic-base';
