+++ projects/eav-ui/src/app/edit/fields/basic/custom-json-editor/custom-json-editor.component.ts
@@ -2,7 +2,8 @@ import { NgClass, NgStyle } from '@angular/common';
 import { Component, computed, inject } from '@angular/core';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import type * as Monaco from 'monaco-editor';
-import { CustomJsonEditor, FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
+import { CustomJsonEditor } from 'projects/edit-types/src/FieldSettings-CustomJsonEditor';
+import { FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
 import { MonacoEditorComponent } from '../../../../monaco-editor/monaco-editor.component';
 import { JsonSchema } from '../../../../monaco-editor/monaco-editor.models';
 import { InputTypeCatalog } from '../../../../shared/fields/input-type-catalog';
