+++ projects/eav-ui/src/app/edit/fields/basic/entity-content-blocks/entity-content-blocks-logic.ts
@@ -1,4 +1,5 @@
-import { FieldSettings, FieldSettingsPickerMerged } from '../../../../../../../edit-types/src/FieldSettings';
+import { FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
+import { FieldSettingsPickerMerged } from '../../../../../../../edit-types/src/FieldSettings-Pickers';
 import { InputTypeCatalog } from '../../../../shared/fields/input-type-catalog';
 import { FieldLogicBase, FieldLogicUpdate } from '../../logic/field-logic-base';
 import { FieldLogicManager } from '../../logic/field-logic-manager';
