+++ projects/eav-ui/src/app/edit/fields/basic/entity-default/entity-default-logic.ts
@@ -1,7 +1,8 @@
+import { FieldSettingsEntity, FieldSettingsSharedCreate } from '../../../../../../../edit-types/src/FieldSettings-Pickers';
 import { InputTypeCatalog } from '../../../../shared/fields/input-type-catalog';
 import { FieldLogicBase, FieldLogicUpdate } from '../../logic/field-logic-base';
 import { PickerLogicShared } from '../../picker/picker-logic-shared';
-import { FieldSettings, FieldSettingsEntity, FieldSettingsSharedCreate } from './../../../../../../../edit-types/src/FieldSettings';
+import { FieldSettings } from './../../../../../../../edit-types/src/FieldSettings';
 
 export class EntityDefaultLogic extends FieldLogicBase {
   name = InputTypeCatalog.EntityDefault;
