+++ projects/eav-ui/src/app/edit/fields/basic/hyperlink-default/hyperlink-default-logic.ts
@@ -1,4 +1,5 @@
-import { FieldSettings, Hyperlink } from '../../../../../../../edit-types/src/FieldSettings';
+import { Hyperlink } from 'projects/edit-types/src/FieldSettings-Hyperlink';
+import { FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
 import { InputTypeCatalog } from '../../../../shared/fields/input-type-catalog';
 import { FieldLogicBase, FieldLogicUpdate } from '../../logic/field-logic-base';
