+++ projects/eav-ui/src/app/edit/fields/basic/string-default/string-default-logic.ts
@@ -1,4 +1,5 @@
-import { FieldSettings, FieldSettingsStringDefault } from '../../../../../../../edit-types/src/FieldSettings';
+import { FieldSettingsStringDefault } from 'projects/edit-types/src/FieldSettings-String';
+import { FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
 import { InputTypeCatalog } from '../../../../shared/fields/input-type-catalog';
 import { FieldLogicBase, FieldLogicUpdate } from '../../logic/field-logic-base';
