+++ projects/eav-ui/src/app/edit/fields/basic/string-dropdown/string-dropdown-logic.ts
@@ -1,5 +1,6 @@
 import { Of } from '../../../../../../../core';
-import { FieldSettings, FieldSettingsPickerMerged, StringDropdown } from '../../../../../../../edit-types/src/FieldSettings';
+import { FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
+import { FieldSettingsPickerMerged, StringDropdown } from '../../../../../../../edit-types/src/FieldSettings-Pickers';
 import { InputTypeCatalog } from '../../../../shared/fields/input-type-catalog';
 import { FieldLogicBase, FieldLogicUpdate } from '../../logic/field-logic-base';
 import { PickerLogicShared } from '../../picker/picker-logic-shared';
