+++ projects/eav-ui/src/app/edit/fields/basic/string-template-picker/string-template-picker.component.ts
@@ -7,9 +7,10 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
 import { MatSelectModule } from '@angular/material/select';
 import { TranslateModule } from '@ngx-translate/core';
+import { StringTemplatePicker } from 'projects/edit-types/src/FieldSettings-String';
 import { take } from 'rxjs';
 import { transient } from '../../../../../../../core/transient';
-import { FieldSettings, StringTemplatePicker } from '../../../../../../../edit-types/src/FieldSettings';
+import { FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
 import { SourceService } from '../../../../code-editor/services/source.service';
 import { CreateFileDialogComponent } from '../../../../create-file-dialog/create-file-dialog.component';
 import { CreateFileDialogData, CreateFileDialogResult } from '../../../../create-file-dialog/create-file-dialog.models';
