+++ projects/eav-ui/src/app/edit/fields/picker/adapters/data-adapter-entity-base.ts
@@ -2,9 +2,9 @@ import { HttpErrorResponse } from "@angular/common/http";
 import { inject } from '@angular/core';
 import { MatSnackBar } from "@angular/material/snack-bar";
 import { TranslateService } from "@ngx-translate/core";
-import { FieldSettingsPickerMerged } from 'projects/edit-types/index-for-documentation';
 import { take } from 'rxjs';
 import { transient } from '../../../../../../../core/transient';
+import { FieldSettingsPickerMerged } from '../../../../../../../edit-types/src/FieldSettings-Pickers';
 import { FieldSettingsWithPickerSource } from '../../../../../../../edit-types/src/PickerSources';
 import { EditForm, EditPrep } from "../../../../../app/shared/models/edit-form.model";
 import { EntityService } from "../../../../../app/shared/services/entity.service";
