+++ projects/eav-ui/src/app/edit/fields/picker/adapters/state-ui-mapper-number-array.ts
@@ -1,4 +1,3 @@
-import { FieldSettings, FieldSettingsSharedSeparator } from '../../../../../../../edit-types/src/FieldSettings';
 import { convertArrayToString } from '../picker.helpers';
 import { StateUiMapperWithSettingsBase } from './state-ui-mapper-base';
 
@@ -23,7 +22,7 @@ export class StateUiMapperNumberArray extends StateUiMapperWithSettingsBase<numb
   toState(uiValue: string[]): number {
     const stringValue = (uiValue == null || uiValue.length === 0)
       ? ''
-      : convertArrayToString(uiValue, (this.settings() as FieldSettings & FieldSettingsSharedSeparator).Separator);
+      : convertArrayToString(uiValue, this.settings().Separator);
 
     const parsedNumber = Number(stringValue);
     return isNaN(parsedNumber) ? null : parsedNumber;
