+++ projects/eav-ui/src/app/edit/fields/picker/adapters/state-ui-mapper-string-array.ts
@@ -1,4 +1,3 @@
-import { FieldSettings, FieldSettingsSharedSeparator } from '../../../../../../../edit-types/src/FieldSettings';
 import { convertArrayToString } from '../picker.helpers';
 import { StateUiMapperWithSettingsBase } from './state-ui-mapper-base';
 
@@ -14,7 +13,7 @@ export class StateUiMapperStringArray extends StateUiMapperWithSettingsBase<stri
   toUi(state: string): string[] {
     return (state == null || state === '')
       ? []
-      : state.split((this.settings() as FieldSettings & FieldSettingsSharedSeparator).Separator) ?? [];
+      : state.split(this.settings().Separator) ?? [];
   }
 
   /**
@@ -23,6 +22,6 @@ export class StateUiMapperStringArray extends StateUiMapperWithSettingsBase<stri
   toState(uiValue: string[]): string {
     return (uiValue == null || uiValue.length === 0) 
       ? ''
-      : convertArrayToString(uiValue, (this.settings() as FieldSettings & FieldSettingsSharedSeparator).Separator);
+      : convertArrayToString(uiValue, this.settings().Separator);
   }
 }
