+++ projects/eav-ui/src/app/edit/fields/picker/picker-tree/picker-tree-data-service.ts
@@ -1,5 +1,6 @@
 import { Injectable, Injector, Signal, effect, runInInjectionContext } from '@angular/core';
-import { FieldSettings, FieldSettingsPicker } from '../../../../../../../edit-types/src/FieldSettings';
+import { FieldSettingsPicker } from 'projects/edit-types/src/FieldSettings-Pickers';
+import { FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
 import { RelationshipParentChild } from '../../../../../../../edit-types/src/PickerModeTree';
 import { classLog } from '../../../../shared/logging';
 import { PickerItem } from '../models/picker-item.model';
