+++ projects/eav-ui/src/app/edit/fields/wrappers/group/collapsible-wrapper-logic.ts
@@ -1,4 +1,5 @@
-import { EmptyDefault, FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
+import { EmptyDefault } from 'projects/edit-types/src/FieldSettings-EmptyDefault';
+import { FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
 import { InputTypeCatalog } from '../../../../shared/fields/input-type-catalog';
 import { FieldLogicBase, FieldLogicUpdate } from '../../logic/field-logic-base';
