+++ projects/eav-ui/src/app/edit/fields/wrappers/group/collapsible-wrapper.component.ts
@@ -2,7 +2,8 @@ import { NgClass } from '@angular/common';
 import { Component, ViewChild, ViewContainerRef, computed, inject } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
-import { EmptyDefault, FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
+import { EmptyDefault } from 'projects/edit-types/src/FieldSettings-EmptyDefault';
+import { FieldSettings } from '../../../../../../../edit-types/src/FieldSettings';
 import { classLog } from '../../../../shared/logging';
 import { SafeHtmlPipe } from '../../../../shared/pipes/safe-html.pipe';
 import { FieldsSettingsService } from '../../../state/fields-settings.service';
