+++ projects/edit-types/src/FieldSettings-Boolean.ts
@@ -1,0 +1,11 @@
+/**
+ * @Boolean
+ */
+export interface FieldSettingsBoolean {
+  TitleTrue: string;
+  TitleFalse: string;
+  TitleIndeterminate: string;
+  ReverseToggle?: boolean;
+  /** Label for Boolean fields, calculated on the fly by the logic */
+  _label: string;
+}
