+++ projects/edit-types/src/FieldSettings-CustomGps.ts
@@ -1,0 +1,8 @@
+/**
+ * @custom-gps
+ */
+export interface CustomGps {
+  LatField: string;
+  LongField: string;
+  AddressMask: string;
+}
