+++ projects/edit-types/src/FieldSettings-CustomJsonEditor.ts
@@ -1,0 +1,12 @@
+/**
+ * @custom-json-editor
+ */
+export interface CustomJsonEditor {
+  Rows: number;
+  JsonValidation: 'strict' | 'light' | 'none';
+  JsonSchemaMode: 'strict' | 'light' | 'none';
+  JsonSchemaSource: 'link' | 'raw';
+  JsonSchemaUrl: string;
+  JsonSchemaRaw: string;
+  JsonCommentsAllowed: boolean;
+}
