+++ projects/edit-types/src/FieldSettings-DateTime.ts
@@ -1,0 +1,7 @@
+
+/**
+ * @DateTime
+ */
+export interface FieldSettingsDateTime {
+  UseTimePicker: boolean;
+}
