+++ projects/edit-types/src/FieldSettings-EmptyDefault.ts
@@ -1,0 +1,11 @@
+/**
+ * @empty-default
+ */
+export interface EmptyDefault {
+  /** @deprecated */
+  DefaultCollapsed: boolean;
+  /** DefaultCollapsed is copied to Collapsed and then deleted  */
+  Collapsed: boolean;
+}
+
+// Note that empty-end / empty-message don't have additional fields
