+++ projects/edit-types/src/FieldSettings-Hyperlink.ts
@@ -1,0 +1,27 @@
+
+/**
+ * @Hyperlink
+ */
+export interface Hyperlink {
+  FileFilter: string;
+  /** @deprecated */
+  DefaultDialog: string;
+  ShowPagePicker: boolean;
+  ShowImageManager: boolean;
+  ShowFileManager: boolean;
+  ShowAdam: boolean;
+  Buttons: string;
+  Paths: string;
+  ServerResourceMapping: string;
+  EnableImageConfiguration: boolean;
+}
+
+/**
+ * @hyperlink-library
+ */
+export interface HyperlinkLibrary extends Hyperlink {
+  FolderDepth: number;
+  AllowAssetsInRoot: boolean;
+  MetadataContentTypes: string;
+}
+
