+++ projects/edit-types/src/FieldSettings-Number.ts
@@ -1,0 +1,11 @@
+/**
+ * @Number
+ */
+export interface FieldSettingsNumber {
+  Decimals: number;
+  Min: number;
+  Max: number;
+  AddressMask: string;
+  ValidationRegEx: string;
+  ValidationRegExJavaScript: string;
+}
