+++ projects/field-string-wysiwyg/src/config/tinymce-configurator.ts
@@ -1,8 +1,8 @@
+import { StringWysiwyg } from 'projects/edit-types/src/FieldSettings-String';
 import type { Editor } from 'tinymce';
 import { classLog } from '../../../../projects/eav-ui/src/app/shared/logging';
 import { AddOnSettings } from '../../../edit-types/src/AddOnSettings';
 import { Connector } from '../../../edit-types/src/Connector';
-import { StringWysiwyg } from '../../../edit-types/src/FieldSettings';
 import { WysiwygReconfigure } from '../../../edit-types/src/WysiwygReconfigure';
 import * as DisplayModes from '../constants/display-modes';
 import * as contentStyle from '../editor/tinymce-content.scss';
